ray_cheng_backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
SQLAlchemy==2.0.30
psycopg[binary]==3.1.19
pydantic==2.7.1
pydantic-settings==2.2.1
python-jose==3.3.0
passlib[bcrypt]==1.7.4
email-validator==2.2.0

